Feature led class

+++ code/headers/hsv_s.hpp
@@ -1,12 +1,14 @@
 #pragma once
 
+#include <stdint.h>
+
 namespace r2d2::led_strip {
     /**
      *  \brief
      *  Struct which contains the hue, saturation and value
      */
     struct hsv_s {
-        uint8_t hue;
+        uint16_t hue;
 	uint8_t saturation;
         uint8_t value;
     };
